Feature(Services): Refactor functions inside structure to services
- Moved functions from structures to services
- Added more tests
