+++ backend/src/services/statistics.test.ts
@@ -1,0 +1,21 @@
+import { mergeCountObjectArrays } from "./statistics";
+
+describe('Satistics', () => {
+  test('mergeCountObjectArray function', () => {
+    const target = [
+      {name: 2, count: 1},
+      {name: 3, count: 2}
+    ]
+    const selected = [
+      {name: 2, count: 1},
+      {name: 4, count: 2}
+    ]
+    const expected = [
+      {name: 2, count: 2},
+      {name: 3, count: 2},
+      {name: 4, count: 2}
+    ]
+    mergeCountObjectArrays(target, selected);
+    expect(target).toEqual(expected);
+  });
+})
