+++ backend/src/services/statistics.ts
